+++ backend/src/main/java/ru/itmo/accontrol/repository/AirConditionerRepository.java
@@ -9,17 +9,17 @@ import ru.itmo.accontrol.entity.AirConditioner;
 import java.util.List;
 
 @Repository
-public interface AirConditionerRepository extends JpaRepository<AirConditioner, Long> {
+public interface AirConditionerRepository extends JpaRepository<AirConditioner, Integer> {
     
-    List<AirConditioner> findByRoomId(Long roomId);
+    List<AirConditioner> findByRoomId(Integer roomId);
     
     List<AirConditioner> findByStatus(String status);
     
     List<AirConditioner> findByMode(String mode);
     
     @Query("SELECT ac FROM AirConditioner ac WHERE ac.room.building.id = :buildingId")
-    List<AirConditioner> findByBuildingId(@Param("buildingId") Long buildingId);
+    List<AirConditioner> findByBuildingId(@Param("buildingId") Integer buildingId);
     
     @Query("SELECT ac FROM AirConditioner ac JOIN ac.groups g WHERE g.id = :groupId")
-    List<AirConditioner> findByGroupId(@Param("groupId") Long groupId);
+    List<AirConditioner> findByGroupId(@Param("groupId") Integer groupId);
 }
